10/07/2024

+++ ImageUploadApi/ImageUploadApi/src/main/java/com/dicap/ImageUploadApi/ImageSevice.java
@@ -1,6 +1,5 @@
 package com.dicap.ImageUploadApi;
 
-import com.zaxxer.hikari.HikariDataSource;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import jakarta.transaction.Transactional;
@@ -29,13 +28,22 @@ public class ImageSevice {
         return imageRepository.count();
     }
 
-
-
     public ImageEntity incrementLikes(Long id) {
         Optional<ImageEntity> optionalImage = imageRepository.findById(id);
         if (optionalImage.isPresent()) {
             ImageEntity image = optionalImage.get();
             image.setLikes(image.getLikes() + 1);
+            return imageRepository.save(image);
+        } else {
+            throw new RuntimeException("Image not found");
+        }
+    }
+
+    public ImageEntity decrementLikes(Long id) {
+        Optional<ImageEntity> optionalImage = imageRepository.findById(id);
+        if (optionalImage.isPresent()) {
+            ImageEntity image = optionalImage.get();
+            image.setLikes(image.getLikes() - 1);
             return imageRepository.save(image);
         } else {
             throw new RuntimeException("Image not found");
@@ -59,8 +67,13 @@ public class ImageSevice {
         return imageRepository.findById(id);
     }
 
-    public void deleteImage(Long id) {
-        imageRepository.deleteById(id);
+    public String deleteImage(Long id) {
+        Optional<ImageEntity> opt = imageRepository.findById(id);
+        if (opt.isPresent()) {
+            imageRepository.deleteById(id);
+            return "Deleted succesfully !";
+        }
+        return "Error while deleting, please try againt !";
     }
 
     public List<ImageEntity> getAllImages() {
@@ -68,13 +81,28 @@ public class ImageSevice {
     }
 
     @Transactional
-    public void deleteAllImages() {
+    public String deleteAllImages() {
         imageRepository.deleteAll();
+        return "All the images was deleted successfully !";
     }
 
-    public List<ImageEntity> getAllByPhotographeId(Long photographeId){return imageRepository.findByPhotographeId(photographeId);}
+    public List<ImageEntity> getAllByPhotographeId(Long photographeId,boolean isDeleted){
+        Session session = entityManager.unwrap(Session.class);
+        Filter filter = session.enableFilter("deleteImageFilter");
+        filter.setParameter("isDeleted", isDeleted);
+        List<ImageEntity> products =  imageRepository.findByPhotographeId(photographeId);
+        session.disableFilter("deletedProductFilter");
+        return products;
+    }
 
-    public List<ImageEntity> getAllByCategorieId(Long categorieId){return imageRepository.findByCategorieId(categorieId);}
+    public List<ImageEntity> getAllByCategorieId(Long categorieId, boolean isDeleted){
+        Session session = entityManager.unwrap(Session.class);
+        Filter filter = session.enableFilter("deleteImageFilter");
+        filter.setParameter("isDeleted", isDeleted);
+        List<ImageEntity> products = imageRepository.findByCategorieId(categorieId);
+        session.disableFilter("deletedProductFilter");
+        return products;
+    }
 
     public List<ImageEntity> getAllById(Long id){
         return imageRepository.findAllById(Collections.singleton(id));
@@ -88,5 +116,16 @@ public class ImageSevice {
         List<ImageEntity> products =  imageRepository.findAll();
         session.disableFilter("deletedProductFilter");
         return products;
+    }
+
+    public ImageEntity restore(Long id) {
+        Optional<ImageEntity> optionalImage = imageRepository.findById(id);
+        if (optionalImage.isPresent()) {
+            ImageEntity image = optionalImage.get();
+            image.setDeleted(Boolean.FALSE);
+            return imageRepository.save(image);
+        } else {
+            throw new RuntimeException("Image not found");
+        }
     }
 }
